requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
opencv-python
numpy
easyocr
translators
deep-translator
torch==2.7.1+cu128
torchvision==0.22.1+cu128
torchaudio==2.7.1+cu128
mysql-connector-python
psutil
netifaces